Safari-609.1.11.1/LayoutTests/streams/shadowing-getReader.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This test should run without throwing an exception.");

var readerConstructor = new ReadableStream().getReader().constructor;

var stream = new ReadableStream();
stream.getReader = "foo";

shouldBeTrue("new readerConstructor(stream) instanceof readerConstructor");

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
