Safari-611.1.10/LayoutTests/streams/shadowing-getReader-expected.txt

This test should run without throwing an exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new readerConstructor(stream) instanceof readerConstructor is true
PASS successfullyParsed is true

TEST COMPLETE
